lots of experimenting

+++ ShoeShockApp/View/CartCell.swift
@@ -52,7 +52,10 @@ class CartCell: UITableViewCell {
         let selectedShoe = SelectedShoe(shoe: shoe, quantity: 1)
         let cartShoes = cartService.cart
         guard let index = cartShoes.firstIndex(of: selectedShoe) else { return }
+        guard let favIndex = dataService.favoritedShoes.firstIndex(of: selectedShoe) else { return }
         if shoeQuantity == 0 {
+            dataService.favoritedShoes[favIndex].isFavorited = false
+            dataService.removeShoeFromFavorites(selectedShoe: selectedShoe)
             cartShoes[index].isFavorited = false
             cartShoes[index].isInCart = false
             cartService.removeShoe(shoe: shoe)
